Add hanging test as a bug reproduction.
This test shows that the earlier commit

    Fix race condition hanging bug in drain phase.

was not enough to fix such cases.

The rationale is as following:

---

Need thread 1 as blocker such that `tryYieldQueueHead` in cruise phase never yields anything.
Then need 3 items of thread 2 in a row to:

1. take its shutdown-Nothing and shut down,
2. ensure that another shutdown-Nothing is put into the `inVar` successfully,
3. ensure that another shutdown-Nothing blocks on being put into the `inVar`.

The thread 1 will also need an associated shutdown-Nothing.
This means we need to generate at least 4 shutdown-Nothings. This means we need at least 4 threads.
To get 3 subsequent items, we need `workerOutputBufferSize` be at least 2.

So we use `concurrentMapM_ 4 2` and construct this scenario
of threads being in order in the `outVar`:

    1   2 2 2   (after this, three 3s and three 4s in arbitrary order)

for example

    1   2 2 2   3 3 3   4 4 4

This can be done by feeding e.g. `[1,   0,0,0,   0,0,0,   0,0,0]`
into the conduit, where the the monadic action sleeps that many
seconds, and inserting sleeps into the code like this:

     forConcurrently_ [1..numThreads] $ \workerNum -> do
    +  threadDelay (workerNum * 100000)

This reproduction works no matter whether in the drain phase

    yieldQueueHead
    putInVar Nothing

or

    putInVar Nothing
    yieldQueueHead

is used.

Inserting such sleep into the code _immediately_ reproduces the
issue on the first try, but in this commit we do not modify the code;
instead the added test case just tries a couple times and usually
succeeds reproducing within the first 20 loop iterations.